fix: bigdicimal로 맞춰주기

+++ src/main/java/com/ebiz/tableorder/menu/dto/SalesDataPoint.java
@@ -3,14 +3,15 @@ package com.ebiz.tableorder.menu.dto;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class SalesDataPoint {
     private final Integer hour;
-    private final Long revenue;
+    private final BigDecimal revenue;   // ← BigDecimal 으로 맞춰줍니다
 
-    // JPQL new 구문과 완전히 일치하도록 반드시 선언해야 합니다.
-    public SalesDataPoint(Integer hour, Long revenue) {
-        this.hour = hour;
+    public SalesDataPoint(Integer hour, BigDecimal revenue) {
+        this.hour    = hour;
         this.revenue = revenue;
     }
 }
